chore: add comment & return typ
