+++ src/internal/TsGenerator/factory/VariableDeclaration.ts
@@ -1,4 +1,5 @@
 import ts from "typescript";
+import { generateComment } from "./utils";
 
 export interface Params {
   name: string | ts.BindingName;
