+++ src/internal/TsGenerator/factory/VariableDeclarationList.ts
@@ -1,4 +1,5 @@
 import ts from "typescript";
+import { generateComment } from "./utils";
 
 const flags = {
   const: ts.NodeFlags.Const,
@@ -7,6 +8,8 @@ const flags = {
 export interface Params {
   declarations: readonly ts.VariableDeclaration[];
   flag: keyof typeof flags;
+  comment?: string;
+  deprecated?: boolean;
 }
 
 export interface Factory {
@@ -17,6 +20,10 @@ export const create =
   ({ factory }: Pick<ts.TransformationContext, "factory">): Factory["create"] =>
   (params: Params): ts.VariableDeclarationList => {
     const node = factory.createVariableDeclarationList(params.declarations, flags[params.flag]);
+    if (params.comment) {
+      const comment = generateComment(params.comment, params.deprecated);
+      return ts.addSyntheticLeadingComment(node, ts.SyntaxKind.MultiLineCommentTrivia, comment.value, comment.hasTrailingNewLine);
+    }
     return node;
   };
